% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_inputs.R
\name{actionElement}
\alias{actionElement}
\title{Make a UI element react to clicks.}
\usage{
actionElement(clickId, el)
}
\arguments{
\item{clickId}{Id of the click handler.}

\item{el}{The \code{shiny.tag} element to modify.}
}
\value{
The modified \code{shiny.tag} element.
}
\description{
Encloses an arbitrary UI element in a div with the *action-button* class.
}
