% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{b64.uid}
\alias{b64.uid}
\title{Generate a random string}
\usage{
b64.uid(size = 64)
}
\arguments{
\item{size}{length of the string to generate.}
}
\value{
a random string.
}
\description{
Generates a random string of specified length using valid
base 64 characters.
}
