% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_generator.R
\name{clear.comments}
\alias{clear.comments}
\title{Clear comments from text}
\usage{
clear.comments(text, n = 1)
}
\arguments{
\item{text}{a string}

\item{n}{number of # symbols (i.e. n = 2 clears text appearing after ## or ###
but not #)}
}
\value{
the clean text
}
\description{
Clear comments from text
}
