% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modal.R
\name{create_file_import_modal}
\alias{create_file_import_modal}
\title{Create a file import modal}
\usage{
create_file_import_modal(id, session)
}
\arguments{
\item{session}{}
}
\value{

}
\description{
Creates a modal.
}
