% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modal.R
\name{create_text_input_modal}
\alias{create_text_input_modal}
\title{Create a modal with a single text input}
\usage{
create_text_input_modal(id, session)
}
\arguments{
\item{session}{}
}
\value{

}
\description{
Create a modal with a single text input
}
