% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_handlers.R
\name{createInputLayout}
\alias{createInputLayout}
\title{Create the UI For the Inputs Panel}
\usage{
createInputLayout(
  pages,
  groups,
  inputs,
  ns = identity,
  inputs.style = c("tabset", "collapses", "flowing")
)
}
\arguments{
\item{pages}{list of pages.}

\item{groups}{list of groups.}

\item{inputs}{list of inputs.}

\item{ns}{the namespace function to use for creating tag ids}

\item{inputs.style}{the style of the input panel. \code{tabset} generates a
tabsetPanel with each page in a tabPanel element. \code{collapses} uses
Bootstrap 4 collapses to compartmentalize pages. \code{flowing} places
all pages vertically in the same container and uses h5 tags for the
page titles.}
}
\value{
a tag.list structure.
}
\description{
Create the UI For the Inputs Panel
}
