% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_inputs.R
\name{fileInput}
\alias{fileInput}
\title{File Upload Control}
\usage{
fileInput(id, label, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Fixes an issue where the page jumps to top when clicking the Browse... button.
}
