% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slate.R
\name{getSlateUserSource}
\alias{getSlateUserSource}
\title{Get formatted user code from slate}
\usage{
getSlateUserSource(x, headers = TRUE)
}
\arguments{
\item{x}{the slate data structure}

\item{headers}{whether to include section headers as comments}
}
\value{
a single character string
}
\description{
Get formatted user code from slate
}
