% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slate.R
\name{getSlateUserSources}
\alias{getSlateUserSources}
\title{Get all user sources from a slate}
\usage{
getSlateUserSources(x)
}
\arguments{
\item{x}{the slate structure}
}
\value{
a list with elements \code{toplevel} and \code{output}, with \code{output} being a list with each
element named after an output defined in the blueprint.
}
\description{
Get user source code for all parts of a slate in the form of a names list of
strings.
}
