% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rslates.R
\name{initServerOptions}
\alias{initServerOptions}
\title{Initialize server options from file}
\usage{
initServerOptions(
  config.file = system.file("rslates.yaml", package = "rslates")
)
}
\arguments{
\item{config.file}{}
}
\value{

}
\description{
Initialize server options from file
}
\examples{
initServerOptions(config.file = "my_settings.yaml")
}
