% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_handlers.R
\name{input_handler}
\alias{input_handler}
\title{Define behaviour of an input type}
\usage{
input_handler(
  create.ui = function(...) {     tagList() },
  update.ui = function(...) { },
  get.value = function(id, input, ...) {     input[[id]] }
)
}
\arguments{
\item{create.ui}{function used to create the input UI.}

\item{update.ui}{function used to update the input UI.}

\item{get.value}{function to get the current value of the input.}
}
\value{
a list of functions to handle the the input type.
}
\description{
Define behaviour of an input type
}
