% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rslates.R
\name{loadBlueprints}
\alias{loadBlueprints}
\title{Load JSON blueprints from file or folder}
\usage{
loadBlueprints(path)
}
\arguments{
\item{path}{}
}
\value{

}
\description{
Load JSON blueprints from file or folder
}
