% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessor.R
\name{preprocessDirective}
\alias{preprocessDirective}
\title{Parse a Preprocessor Directive}
\usage{
preprocessDirective(text, type = NULL)
}
\arguments{
\item{text}{Text of the full directive including all parameters.}

\item{type}{Type of directive if the \verb{$@<directive>} has been scrubbed.}
}
\value{
The resulting blueprint element structure obtained from a call to \code{slateInput},
\code{slateGroup}, \code{slatePage}, \code{slateImport} or \code{slateExport}.
}
\description{
Parse a preprocessor directive of the form \verb{$@<directive> name, ...}.
}
