% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessor.R
\name{preprocessInlineInputDirective}
\alias{preprocessInlineInputDirective}
\title{Preprocess an Inline Input Directive}
\usage{
preprocessInlineInputDirective(text)
}
\arguments{
\item{text}{The directive text.}
}
\value{
An input obtained from a call to \code{slateInput}.
}
\description{
Preprocess an Inline Input Directive
}
\details{
Preprocess a directive of the form \code{name(parameter1, parameter2, ...)}. We do this
by transforming the text into the form \verb{name, parameter1, parameter2, ...} and passing
it on to \code{preprocessDirective()}.
}
