% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_generator.R
\name{preprocessInputs}
\alias{preprocessInputs}
\title{Extract Input Definitions From Blueprint Source}
\usage{
preprocessInputs(text)
}
\arguments{
\item{text}{a single character string representing the source code
to preprocess.}
}
\value{
A list where each element represents an input. Each element of
an input represents a option that can be passed on to slateInput() in
order to construct the input.
}
\description{
Extract Input Definitions From Blueprint Source
}
