% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessor.R
\name{preprocessSubstitutionOptions}
\alias{preprocessSubstitutionOptions}
\title{Preprocess Substitution Options for a Variable or Block}
\usage{
preprocessSubstitutionOptions(text)
}
\arguments{
\item{text}{the text of the options}
}
\value{
A character vector of individual options
}
\description{
Preprocess Substitution Options for a Variable or Block
}
\details{
throws an error if any of the options is invalid.
}
