% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_editor.R
\name{projectEditorUI}
\alias{projectEditorUI}
\title{Editor Panel UI Function}
\usage{
projectEditorUI(id, project)
}
\arguments{
\item{id}{the panel id.}
}
\value{
The HTML representing the panel.
}
\description{
Editor Panel UI Function
}
