% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessor.R
\name{resolveSubstitutionOptions}
\alias{resolveSubstitutionOptions}
\title{Resolves Conflicts Between Variable and Block Substitution Options}
\usage{
resolveSubstitutionOptions(var.opts, block.opts)
}
\arguments{
\item{var.opts}{character vector of variable options}

\item{block.opts}{character vector of block options}
}
\value{
The resolved variable options.
}
\description{
Resolves Conflicts Between Variable and Block Substitution Options
}
\details{
throws an error if the options combination cannot be resolved.
}
