% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprint.R
\name{restoreBlueprint}
\alias{restoreBlueprint}
\title{Restore a blueprint by filling-in default values}
\usage{
restoreBlueprint(blueprint)
}
\arguments{
\item{blueprint}{the blueprint to be restored.}
}
\value{
the restored blueprint containing all fields, including those with default values.
}
\description{
Restore a blueprint by filling-in default values
}
\seealso{
\code{\link[=simplifyBlueprint]{simplifyBlueprint()}} to remove default values from a blueprint.
}
