% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{seq.uid}
\alias{seq.uid}
\title{Generate a sequencial id}
\usage{
\method{seq}{uid}(prefix = "_")
}
\arguments{
\item{prefix}{the prefix string to use.}
}
\value{
a unique string id.
}
\description{
This function generates a unique sequencial identifier with a given prefix
each time it is called in the current session.
}
