% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprint.R
\name{simplifyBlueprint}
\alias{simplifyBlueprint}
\title{Simplify a blueprint by removing default values}
\usage{
simplifyBlueprint(blueprint)
}
\arguments{
\item{blueprint}{the blueprint to simplify.}
}
\value{
The simplified blueprint. This blueprint is ideal for export in JSON format or other format.
}
\description{
Simplify a blueprint by removing default values
}
\seealso{
[restoreBlueprint()] to restore a simplified blueprint to its initial state.
}
