% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprint.R
\name{slateGroup}
\alias{slateGroup}
\title{Define a Group For Slate Inputs}
\usage{
slateGroup(
  name,
  ...,
  parent = "auto",
  description = "",
  layout = "flow-2",
  condition = NULL
)
}
\arguments{
\item{condition}{}
}
\value{

}
\description{
Define a Group For Slate Inputs
}
