% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprint.R
\name{slateImport}
\alias{slateImport}
\title{Define a slate import}
\usage{
slateImport(name, type, value = "", description = "")
}
\arguments{
\item{type}{file or dataframe}

\item{value}{path of file or name of dataframe}
}
\value{

}
\description{
Define a slate import
}
