% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprint.R
\name{slateInput}
\alias{slateInput}
\title{Define a Slate Input}
\usage{
slateInput(
  name,
  type = "character",
  default = NULL,
  allow.null = FALSE,
  parent = "auto",
  long.name = NULL,
  description = "",
  condition = NULL,
  wizards = list(),
  ...
)
}
\arguments{
\item{name}{the name of the input.}

\item{type}{the type of the input. use \code{names(input.handlers)} for a list of valid
input types.}

\item{default}{the default value for the input.}

\item{parent}{the parent of this input. Can be either the name of a group or page, or NULL
to have the input be a child of root, or "auto" (the default) to infer the parent based
on the order of page, group and input definitions. See \code{\link[=inferSlateLayout]{inferSlateLayout()}} for details.}

\item{long.name}{the name to display in the user interface, or NULL to use the input's name}

\item{description}{the description of the input. Usually this is shown via a tooltip.}

\item{condition}{an expression that determines the visibility of this input, or NULL to
always show the input.}

\item{wizards}{a list of wizards associated with this input. Wizards will ususally be
displayed in a dropdown menu above the input. See \code{\link[=slateWizard]{slateWizard()}} for details.}

\item{...}{additional input parameters to be stored in the input structure. These are usually
type-specific parameters, such as \code{choices} if the input is of type \code{select}.}

\item{input.type}{\emph{legacy: use \code{type} instead}}

\item{null}{if TRUE, a toggle is displayed to allow the user to select NULL as the input value.}
}
\value{
A list of input parameters suitable to be passed along to other \code{rslates} functions.
}
\description{
Define a Slate Input
}
