% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprint.R
\name{slatePage}
\alias{slatePage}
\title{Define a Page For Slate Inputs}
\usage{
slatePage(name, title = name, layout = "flow-2", description = "")
}
\arguments{
\item{description}{}
}
\value{

}
\description{
Define a Page For Slate Inputs
}
