% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_editor.R
\name{slates_editServer}
\alias{slates_editServer}
\title{Editor Panel Server Function}
\usage{
slates_editServer(input, output, session, project, session.data)
}
\description{
Editor Panel Server Function
}
