% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_editor.R
\name{slates_editUI}
\alias{slates_editUI}
\title{Editor Panel UI Function}
\usage{
slates_editUI(id, project.data)
}
\arguments{
\item{id}{the panel id.}
}
\value{
The HTML representing the panel.
}
\description{
Editor Panel UI Function
}
