% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_inputs.R
\name{slatesCollapsePanel}
\alias{slatesCollapsePanel}
\title{Create a bsCollapsePanel with toggle icon.}
\usage{
slatesCollapsePanel(title, ..., value = title, style = NULL)
}
\arguments{
\item{style}{}
}
\value{

}
\description{
Create a bsCollapsePanel with toggle icon.
}
