% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_inputs.R
\name{slatesFileInput}
\alias{slatesFileInput}
\title{Shiny File Upload Control}
\usage{
slatesFileInput(id, label, ..., class = "")
}
\description{
Fixes an issue where the page jumps to top when clicking the "Browse..." button.
}
