% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modal.R
\name{slatesModal}
\alias{slatesModal}
\title{Create a Modal Dialog}
\usage{
slatesModal(
  id,
  session,
  submit.fun,
  ui.fun,
  observers = list(),
  focus.on.show = NULL
)
}
\arguments{
\item{id}{id of the modal dialog.}

\item{session}{server session.}

\item{submit.fun}{a function without arguments that returns the a list
of results of the modal dialog when the user presses OK.}

\item{ui.fun}{a function taking any number of arguments that returns
the ui of the body of the modal.}

\item{observers}{an optional list of observers to be destroyed when the
modal is destroyed.}

\item{focus.on.show}{unused}
}
\value{
a list with the following members. \code{id}: the id of the modal,
\code{show}: a function with signature \verb{function(callback, title = NULL, ...)}
that brings up the modal dialog when called. The \code{callback} function is
called by the modal dialog when the user presses OK. Each value returned
by \code{submit.fun} is passed to this function as an argument. Extra \code{...}
parameters are passed to the \code{ui.fun} function.
}
\description{
Create a Modal Dialog
}
