% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{uiReady}
\alias{uiReady}
\title{Observe shiny inputs for readyness}
\usage{
uiReady(session)
}
\arguments{
\item{session}{the server session}
}
\value{
a reactiveVal that if TRUE indicates the UI has been initialized.
}
\description{
Observe shiny inputs for readyness
}
