% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modal.R
\name{create_new_project_modal}
\alias{create_new_project_modal}
\title{Create the new project modal}
\usage{
create_new_project_modal(id, session)
}
\arguments{
\item{id}{the id for this modal.}

\item{session}{the session from the server function.}
}
\value{
A list containing the \code{show} function used to display the modal after being
created.
}
\description{
Creates a modal to be displayed when creating a new project. Allows
input of the project's initital configuration, such as title and author.
}
