% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slate.R
\name{getSlateUserSources}
\alias{getSlateUserSources}
\title{Get all user sources from a slate}
\usage{
getSlateUserSources(x)
}
\arguments{
\item{x}{the slate data structure}
}
\value{
a list with elements \code{imports}, \code{toplevel} and \code{outputs}.
}
\description{
Get user source code for all parts of a slate in the form of a names list of
strings.
}
