% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_handlers.R
\name{inputHandler}
\alias{inputHandler}
\title{Define behaviour of an input type}
\usage{
inputHandler(
  default.value,
  params.list = list(),
  create.ui = function(...) {     tagList() },
  update.ui = function(...) { },
  get.inputs = function(x, session) {     session$input[[x$id]] },
  get.value = function(x, session = NULL, value = NULL) {     if (is.null(value))      
       value <- session$input[[x$id]]     return(as.character(value)) },
  get.source = function(x, session = NULL, value = NULL) {    
    input.handlers[[x$input.type]]$get.value(x, session, value) },
  observer = function(x, session) { }
)
}
\arguments{
\item{params.list}{list of parameters used by the input type. Each element of the list
describes a parameter and has the following members: name (pretty name of the parameter),
choices (vector of allowed values), default (default value when not specified).}

\item{create.ui}{function used to create the input UI.}

\item{update.ui}{function used to update the input UI.}

\item{get.value}{function that transforms the current value of the input
to the appropriate output string value. If value is provided, then
this value is used instead of the input element value.}
}
\value{
a list of functions to handle the the input type.
}
\description{
This function prepares a structure that defines all aspects of an
  input type: UI creation, UI updating, observer function and string representation
  of the input value.
}
