% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_handlers.R
\name{inputHandler}
\alias{inputHandler}
\title{Define behaviour of an input type}
\usage{
inputHandler(
  create.ui = function(...) {     tagList() },
  update.ui = function(...) { },
  get.value = function(x, session = NULL, value = NULL) {     if (is.null(value))      
       value <- session$input[[x$id]]     return(value) },
  create.observer = function(...) { }
)
}
\arguments{
\item{create.ui}{function used to create the input UI.}

\item{update.ui}{function used to update the input UI.}

\item{get.value}{function that transforms the current value of the input
to the appropriate output string value. If value is provided, then
this value is used instead of the input element value.}
}
\value{
a list of functions to handle the the input type.
}
\description{
Define behaviour of an input type
}
