% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modal.R
\name{slatesModal}
\alias{slatesModal}
\title{Create a Modal Dialog}
\usage{
slatesModal(id, session, submit.fun, ui.fun, observers = list())
}
\arguments{
\item{id}{id of the modal dialog.}

\item{session}{server session.}

\item{submit.fun}{a function without arguments that returns the a list
of results of the modal dialog when the user presses OK.}

\item{ui.fun}{a function taking any number of arguments that returns
the ui of the body of the modal.}

\item{observers}{an optional list of observers to be destroyed when the
modal is destroyed.}
}
\value{
a list with the following members. `id`: the id of the modal,
  `show`: a function with signature `function(callback, title = NULL, ...)`
  that brings up the modal dialog when called. The `callback` function is
  called by the modal dialog when the user presses OK. Each value returned
  by `submit.fun` is passed to this function as an argument. Extra `...`
  parameters are passed to the `ui.fun` function.
}
\description{
Create a Modal Dialog
}
